+++ src/apple/xcode-manager.ts
@@ -164,6 +164,7 @@ function addUploadSymbolsScript(
     }
   }
 
+  const isHomebrewInstalled = fs.existsSync('/opt/homebrew/bin/sentry-cli');
   xcodeProject.addBuildPhase(
     [],
     'PBXShellScriptBuildPhase',
@@ -178,6 +179,7 @@ function addUploadSymbolsScript(
         sentryProject.organization.slug,
         sentryProject.slug,
         uploadSource,
+        isHomebrewInstalled,
       ),
     },
   );